move gradate color function
