+++ frontend/src/scenes/retention/RetentionTable.tsx
@@ -3,9 +3,8 @@ import './RetentionTable.scss'
 import clsx from 'clsx'
 import { mean } from 'd3'
 import { useActions, useValues } from 'kea'
-import { BRAND_BLUE_HSL, gradateColor, PURPLE } from 'lib/colors'
 import { Tooltip } from 'lib/lemon-ui/Tooltip'
-import { range } from 'lib/utils'
+import { BRAND_BLUE_HSL, gradateColor, range } from 'lib/utils'
 import { insightLogic } from 'scenes/insights/insightLogic'
 import { insightVizDataLogic } from 'scenes/insights/insightVizDataLogic'
